+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/BubbleGraph/WeightColorBasedGraph.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import BubbleSingleDataset from '@cerner/terra-graphs/lib/components/Bubble/BubbleSingleDataset';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib//components/Graph.module.scss';
+import '@cerner/terra-graphs/lib//components/Dev.module.scss';
+import getSimpleAxisData from '@cerner/terra-graphs-docs/src/example-datasets/graphConfigObjects/Bubble/simpleAxisData';
+import bubbleDataColorWeightBased from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDataColorWeightBased.js';
+
+/*
+Please refer to the documentation below to see graphConfig and data objects.
+*/
+const graphConfig = utils.deepClone(getSimpleAxisData('#colorBasedBubbleGraph'));
+const dataset = utils.deepClone(bubbleDataColorWeightBased);
+
+export default () => <BubbleSingleDataset graphID="colorBasedBubbleGraph" graphConfig={graphConfig} dataset={[dataset]} />;