+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/BubbleGraph/ColorBasedGraph.test.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import BubbleSingleDataset from '@cerner/terra-graphs/lib/components/Bubble/BubbleSingleDataset';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Dev.module.scss';
+import getSimpleAxisData from '@cerner/terra-graphs-docs/src/example-datasets/graphConfigObjects/Bubble/simpleAxisData';
+import bubbleDataColorBased from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDataColorBased.js';
+
+const graphConfig = utils.deepClone(getSimpleAxisData('#colorBasedBubbleGraph'));
+const dataset = utils.deepClone(bubbleDataColorBased);
+
+export default () => <BubbleSingleDataset graphID="colorBasedBubbleGraph" graphConfig={graphConfig} dataset={[dataset]} />;