+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/BubbleGraph/SingleDataset.test.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import BubbleSingleDataset from '@cerner/terra-graphs/lib/components/Bubble/BubbleSingleDataset';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib//components/Graph.module.scss';
+import '@cerner/terra-graphs/lib//components/Dev.module.scss';
+import getSimpleAxisData from '@cerner/terra-graphs-docs/src/example-datasets/graphConfigObjects/Bubble/SimpleAxisData';
+import bubbleDataBasic from '@cerner/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDataBasic.js';
+
+const graphConfig = utils.deepClone(getSimpleAxisData('#simpleBubbleGraph'));
+const dataset = utils.deepClone(bubbleDataBasic);
+
+export default () => <BubbleSingleDataset graphID="simpleBubbleGraph" graphConfig={graphConfig} dataset={[dataset]} />;