+++ packages/terra-graphs/CHANGELOG.md
@@ -6,6 +6,7 @@
   * Create Components, examples and WDIO test cases for Line, Bar and Paired Result graph types.
   
 * Added
+  * Added bubble components, wdio tests and corresponding examples.
   * Added documentation for `reflowMultipleDatasets`.
 
 * Initial stable release